First commit for this very exciting branch
This branch is not meant for upstream merging/pull requests.
It is meant as an active dev branch where quick hacks can be posted,
that will hopefully be cleaned/tested and made into their own branches for
upstream merges later.

Changes thus far are as follows:

 - Make status report optional.
	dispatch.cpp
	dispatch_ui.cpp
	dispatch_options.h
 - Make it so that QEMU terminates automatically when VMNode (parent)
terminates.
	runtime-dump.cpp
 - Base start timer for timeout on when dispatch receives guest data,
and not first trace (improve accuracy, fault tolerance).
	runtime-dump.cpp
 - Workaround for trace-tag feature assertion when using
   crete_make_concolic in the harness.
	test_pool.cpp